refactor: manage controller position hierarchies through a single service
