+++ app/Filament/Resources/PrenoteResource/RelationManagers/ControllersRelationManager.php
@@ -4,7 +4,7 @@ namespace App\Filament\Resources\PrenoteResource\RelationManagers;
 
 use App\Helpers\Controller\FrequencyFormatter;
 use App\Models\Controller\ControllerPosition;
-use App\Services\PrenoteService;
+use App\Services\ControllerPositionHierarchyService;
 use Filament\Forms;
 use Filament\Resources\Form;
 use Filament\Resources\RelationManagers\RelationManager;
@@ -36,11 +36,11 @@ class ControllersRelationManager extends RelationManager
                     ->label(__('table.prenotes.controller_positions.columns.callsign.label')),
                 Tables\Columns\TextColumn::make('frequency')
                     ->label(__('table.prenotes.controller_positions.columns.frequency.label'))
-                    ->formatStateUsing(fn (float $state) => FrequencyFormatter::formatFrequency($state)),
+                    ->formatStateUsing(fn(float $state) => FrequencyFormatter::formatFrequency($state)),
             ])
             ->headerActions([
                 Tables\Actions\AttachAction::make()
-                    ->form(fn (Tables\Actions\AttachAction $action, ControllersRelationManager $livewire) => [
+                    ->form(fn(Tables\Actions\AttachAction $action, ControllersRelationManager $livewire) => [
                         $action->getRecordSelect(),
                         Forms\Components\Select::make('insert_after')
                             ->label(__('table.prenotes.controller_positions.attach_form.insert_after.label'))
@@ -51,12 +51,12 @@ class ControllersRelationManager extends RelationManager
                                 $livewire->getOwnerRecord()
                                     ->controllers
                                     ->mapWithKeys(
-                                        fn (ControllerPosition $controller) => [$controller->id => $controller->callsign]
+                                        fn(ControllerPosition $controller) => [$controller->id => $controller->callsign]
                                     )
                             ),
                     ])
                     ->using(function (ControllersRelationManager $livewire, $data) {
-                        PrenoteService::insertPositionIntoPrenoteOrder(
+                        ControllerPositionHierarchyService::insertPositionIntoHierarchy(
                             $livewire->getOwnerRecord(),
                             ControllerPosition::findOrFail($data['recordId']),
                             after: isset($data['insert_after'])
@@ -72,7 +72,7 @@ class ControllersRelationManager extends RelationManager
             ->actions([
                 Tables\Actions\Action::make('moveUp')
                     ->action(function (ControllerPosition $record) {
-                        PrenoteService::moveControllerInPrenoteOrder(
+                        ControllerPositionHierarchyService::moveControllerInHierarchy(
                             $record->pivot->pivotParent,
                             $record,
                             true
@@ -80,10 +80,10 @@ class ControllersRelationManager extends RelationManager
                     })
                     ->label(__('table.prenotes.controller_positions.move_up_action.label'))
                     ->icon('heroicon-o-arrow-up')
-                    ->authorize(fn (ControllersRelationManager $livewire) => $livewire->can('moveUp')),
+                    ->authorize(fn(ControllersRelationManager $livewire) => $livewire->can('moveUp')),
                 Tables\Actions\Action::make('moveDown')
                     ->action(function (ControllerPosition $record) {
-                        PrenoteService::moveControllerInPrenoteOrder(
+                        ControllerPositionHierarchyService::moveControllerInHierarchy(
                             $record->pivot->pivotParent,
                             $record,
                             false
@@ -91,10 +91,10 @@ class ControllersRelationManager extends RelationManager
                     })
                     ->label(__('table.prenotes.controller_positions.move_down_action.label'))
                     ->icon('heroicon-o-arrow-down')
-                    ->authorize(fn (ControllersRelationManager $livewire) => $livewire->can('moveUp')),
+                    ->authorize(fn(ControllersRelationManager $livewire) => $livewire->can('moveUp')),
                 Tables\Actions\DetachAction::make()
                     ->using(function (ControllerPosition $record) {
-                        PrenoteService::removeFromPrenoteOrder(
+                        ControllerPositionHierarchyService::removeFromHierarchy(
                             $record->pivot->pivotParent,
                             $record
                         );
