+++ app/Models/Controller/HasControllerHierarchy.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models\Controller;
+
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+
+interface HasControllerHierarchy
+{
+    public function controllers(): BelongsToMany;
+}
