+++ tests/app/Services/HandoffServiceTest.php
@@ -4,16 +4,11 @@ namespace App\Services;
 
 use App\BaseFunctionalTestCase;
 use App\Models\Controller\ControllerPosition;
-use App\Models\Controller\Handoff;
 use Illuminate\Support\Facades\DB;
-use InvalidArgumentException;
 
 class HandoffServiceTest extends BaseFunctionalTestCase
 {
-    /**
-     * @var HandoffService
-     */
-    private $service;
+    private readonly HandoffService $service;
 
     public function setUp(): void
     {
@@ -58,7 +53,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             []
         );
         $this->assertDatabaseHas(
-           'handoffs',
+            'handoffs',
             [
                 'id' => $handoff->id,
                 'description' => 'test',
@@ -77,14 +72,14 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoffs',
+            'handoffs',
             [
                 'id' => $handoff->id,
                 'description' => 'test',
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => $handoff->id,
                 'controller_position_id' => 1,
@@ -92,377 +87,11 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => $handoff->id,
                 'controller_position_id' => 2,
                 'order' => 2,
-            ]
-        );
-    }
-
-    public function testItSetsControllerPositionsForHandoffOrderByControllerId()
-    {
-        HandoffService::setPositionsForHandoffByControllerId(
-            Handoff::findOrFail(1),
-            [
-                2,
-                4,
-                1,
-                3,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 3,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 4,
-            ]
-        );
-    }
-
-    public function testItSetsControllerPositionsForHandoffOrderByControllerCallsign()
-    {
-        HandoffService::setPositionsForHandoffByControllerCallsign(
-            Handoff::findOrFail(1),
-            [
-                'EGLL_N_APP',
-                'LON_C_CTR',
-                'EGLL_S_TWR',
-                'LON_S_CTR',
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 3,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 4,
-            ]
-        );
-    }
-
-    public function testItSetsControllerPositionsForHandoffOrderByController()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(4),
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 3,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 4,
-            ]
-        );
-    }
-
-    public function testItAddsAPositionToTheEndOfTheOrder()
-    {
-        HandoffService::insertPositionIntoHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(4)
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItAddsAPositionToOrderBeforeAnother()
-    {
-        HandoffService::insertPositionIntoHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(4),
-            before: ControllerPosition::findOrFail(2)
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItAddsAPositionToOrderAfterAnother()
-    {
-        HandoffService::insertPositionIntoHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(4),
-            after: ControllerPosition::findOrFail(1)
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItThrowsExceptionIfInsertingBeforePositionNotInOrder()
-    {
-        $this->expectException(InvalidArgumentException::class);
-        HandoffService::insertPositionIntoHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(4),
-            before: ControllerPosition::findOrFail(3)
-        );
-    }
-
-    public function testItRemovesControllersFromAHandoffOrderByModel()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(4),
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::removeFromHandoffOrder(Handoff::findOrFail(1), ControllerPosition::findOrFail(1));
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItRemovesControllersFromAHandoffOrderById()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(4),
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::removeFromHandoffOrder(Handoff::findOrFail(1), 1);
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItRemovesControllersFromAHandoffOrderByCallsign()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(4),
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::removeFromHandoffOrder(Handoff::findOrFail(1), 'EGLL_S_TWR');
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 4,
-                'order' => 2,
-            ]
-        );
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 3,
             ]
         );
     }
@@ -476,7 +105,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 1,
@@ -484,7 +113,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 4,
@@ -492,7 +121,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 2,
@@ -501,7 +130,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 4,
@@ -509,7 +138,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 2,
@@ -517,7 +146,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 3,
@@ -535,7 +164,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 1,
@@ -543,7 +172,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 2,
@@ -551,7 +180,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 4,
@@ -560,7 +189,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 2,
@@ -568,7 +197,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 4,
@@ -576,7 +205,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 3,
@@ -594,7 +223,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 1,
@@ -602,7 +231,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 4,
@@ -610,7 +239,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 2,
@@ -619,7 +248,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 4,
@@ -627,7 +256,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 2,
@@ -635,7 +264,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
             ]
         );
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 3,
@@ -649,7 +278,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         HandoffService::removePositionFromAllHandoffs(ControllerPosition::find(2));
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 1,
@@ -658,7 +287,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 3,
@@ -672,7 +301,7 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         HandoffService::removePositionFromAllHandoffs('EGLL_N_APP');
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 1,
                 'controller_position_id' => 1,
@@ -681,254 +310,12 @@ class HandoffServiceTest extends BaseFunctionalTestCase
         );
 
         $this->assertDatabaseHas(
-           'handoff_orders',
+            'handoff_orders',
             [
                 'handoff_id' => 2,
                 'controller_position_id' => 3,
                 'order' => 1,
             ]
-        );
-    }
-
-    public function testItMovesPositionUpAtStartOfOrder()
-    {
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(1),
-            true
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 2,
-            ]
-        );
-    }
-
-    public function testItMovesPositionUpIfOnlyPosition()
-    {
-        HandoffService::setPositionsForHandoffByController(Handoff::findOrFail(1), [ControllerPosition::findOrFail(1)]);
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(1),
-            true
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-    }
-
-    public function testItMovesPositionUpInTheOrder()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(2),
-            true
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 1,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 2,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItMovesPositionDownAtEndOfOrder()
-    {
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(2),
-            false
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 2,
-            ]
-        );
-    }
-
-    public function testItMovesPositionDownIfOnlyPosition()
-    {
-        HandoffService::setPositionsForHandoffByController(Handoff::findOrFail(1), [ControllerPosition::findOrFail(1)]);
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(1),
-            false
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-    }
-
-    public function testItMovesPositionDownInTheOrder()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::moveControllerInHandoffOrder(
-            Handoff::findOrFail(1),
-            ControllerPosition::findOrFail(2),
-            false
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 2,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItMovesPositionInOrderByIds()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        HandoffService::moveControllerInHandoffOrder(
-            1,
-            2,
-            false
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 1,
-                'order' => 1,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 3,
-                'order' => 2,
-            ]
-        );
-
-        $this->assertDatabaseHas(
-           'handoff_orders',
-            [
-                'handoff_id' => 1,
-                'controller_position_id' => 2,
-                'order' => 3,
-            ]
-        );
-    }
-
-    public function testItThrowsExceptionIfPositionToBeMovedIsNotInOrer()
-    {
-        HandoffService::setPositionsForHandoffByController(
-            Handoff::findOrFail(1),
-            [
-                ControllerPosition::findOrFail(1),
-                ControllerPosition::findOrFail(2),
-                ControllerPosition::findOrFail(3),
-            ]
-        );
-        $this->expectException(InvalidArgumentException::class);
-        HandoffService::moveControllerInHandoffOrder(
-            1,
-            4,
-            false
         );
     }
 }
